% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{calCor_rt}
\alias{calCor_rt}
\title{calCor_rt}
\usage{
calCor_rt(chrDf1, chrDf2)
}
\arguments{
\item{chrDf1}{chrDf1}

\item{chrDf2}{chrDf2}
}
\value{
A double .
}
\description{
Calculate the retention time correlation of two chrDf
}
\examples{
load("./test_data/swath_data.RData")
mchrDfList <- get_chrDfList(ndata = swath_data, m = 34)
calCor_rt(chrDf1 = mchrDfList$ms1[[1]], chrDf2 = mchrDfList$ms2[[5]])
}
