% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{calCor_shape}
\alias{calCor_shape}
\title{calCor_shape}
\usage{
calCor_shape(chrDf1, chrDf2)
}
\arguments{
\item{chrDf1}{chrDf1}

\item{chrDf2}{chrDf2}
}
\value{
A double.
}
\description{
Calculate the peak correlation of two chrDf
}
\examples{
calCor_shape(chrDf1 = mchrDfList$ms1[[1]], chrDf2 = mchrDfList$ms2[[5]])
}
