% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{chrDfList2spectra}
\alias{chrDfList2spectra}
\title{chrDfList2spectra}
\usage{
chrDfList2spectra(chrDfList, ndata)
}
\arguments{
\item{chrDfList}{chrDfList}

\item{ndata}{nth sample data.}
}
\value{
A Spectra object.
}
\description{
Transfor chrDfList to a Spectra object
}
\examples{
chrDfList2spectra(chrDfList = mchrDfList_new, ndata = swath_data)
}
