% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{Deconv4ndata}
\alias{Deconv4ndata}
\title{Deconv4ndata}
\usage{
Deconv4ndata(ndata, weight_rt = 0.5, weight_shape = 0.5, st = 0.7, thread = 1)
}
\arguments{
\item{ndata}{nth sample data.}

\item{weight_rt}{weight_rt.}

\item{weight_shape}{weight_shape.}

\item{st}{socre threshold.}

\item{thread}{prarllel thread.}
}
\value{
A new featureTable_ms1 with ms2 spectra.
}
\description{
Deconv function for nth sample.
}
\examples{
Deconv4ndata(ndata = swath_data)
}
