% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peaks.R
\name{extract_ms1Peaks}
\alias{extract_ms1Peaks}
\alias{extract_ms2Peaks}
\title{Extract ms1 or ms2 peaks}
\usage{
extract_ms1Peaks(
  targetMz,
  dps_ms1,
  scanIndex_ms1,
  rtime_ms1,
  ppm = 5,
  peakwidth = c(5, 20),
  snthresh = 3,
  noise = 100,
  r2thresh = 0.8,
  peakWidthExtend = 1,
  fill_missing = FALSE
)

extract_ms2Peaks(
  ms1PeakDT,
  fragmentMz,
  dps_ms2,
  scanIndex_ms2,
  rtime_ms2,
  ppm = 5,
  peakwidth = c(5, 20),
  snthresh = 3,
  noise = 100,
  r2thresh = 0.8,
  fill_missing = FALSE
)
}
\arguments{
\item{targetMz}{target m/z}

\item{dps_ms1}{\code{data.table()} with mz, intensity, scanIndex}

\item{scanIndex_ms1}{scanIndex of ms1 spectra}

\item{rtime_ms1}{retention time of ms1 spectra}

\item{ppm}{ppm}

\item{peakwidth}{peakwidth}

\item{snthresh}{snthresh}

\item{noise}{noise}

\item{r2thresh}{r2thresh}

\item{peakWidthExtend}{peakWidthExtend}

\item{fill_missing}{whether to fill in missing values in chromatogram}

\item{ms1PeakDT}{\code{data.table()}, one row of ms1PeaksDT}

\item{fragmentMz}{\code{numeric()}, fragment m/z of ms1 peak}

\item{dps_ms2}{dps_ms2 \code{data.table()} with mz, intensity and scanIndex. Data points of ms2 spectra.}

\item{scanIndex_ms2}{scanIndex_ms2 scanIndex of ms2 spectra}

\item{rtime_ms2}{rtime_ms2 retention time of ms2 spectra}
}
\value{
ms1PeaksDT \code{data.table()} with mz, mzmin, mzmax, rt, rtmin, rtmax, into, intb, maxo, r2, intensity(list), rtime(list)

ms2PeaksDT \code{data.table()} with mz, mzmin, mzmax, rt, rtmin, rtmax, maxo, intensity(list), rtime(list)
}
\description{
Extract ms1 peaks from target mz or extract ms2 peaks from ms1PeaksDT
}
\examples{
ms1PeaksDT <- extract_ms1Peaks()
}
