% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_signals.R
\name{fill_missing_signals}
\alias{fill_missing_signals}
\title{Fill missing signals for chromatographic peak}
\usage{
fill_missing_signals(intensity, gap = 1, noise = 100)
}
\arguments{
\item{intensity}{\code{numeric()}, intensity vector}

\item{gap}{\code{integer(1)}, number of missing signals allowed in a chromatographic peak}

\item{noise}{\code{numeric(1)}, noise value of chromatogram}
}
\value{
\code{numeric()}, filled intensity vector
}
\description{
Use approx function to fill missing signals for chromatographic peak
}
