% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrDf.R
\name{filterChrDf}
\alias{filterChrDf}
\title{filterChrDf}
\usage{
filterChrDf(chrDfList, weight_rt = 0.5, weight_shape = 0.5, st = 0.8)
}
\arguments{
\item{chrDfList}{A chrDfList contains ms1 and ms2 chrDf}
}
\value{
A new chrDfList.
}
\description{
Filter less relevant ms2 features
}
\examples{
load("./test_data/swath_data.RData")
chromPeakTable <- dplyr::as_tibble(cbind(xcms::chromPeaks(swath_data),
                                        xcms::chromPeakData(swath_data)),
                                  rownames = "cpid")
chromPeakTable_ms1 <- chromPeakTable \%>\%
 dplyr::filter(ms_level == 1)
chromPeakTable_ms2 <- chromPeakTable \%>\%
 dplyr::filter(ms_level == 2)
# filter chromPeakTable
chromPeakTable_ms1 <- chromPeakTable_ms1 \%>\%
  dplyr::filter(maxo >= 1000)
chromPeakTable_ms2 <- chromPeakTable_ms2 \%>\%
  dplyr::filter(maxo > 100)
mchrDfList <- generate_chrDfList(ndata = swath_data, cpid = "CP34",
                                chromPeakTable_ms1 = chromPeakTable_ms1, chromPeakTable_ms2 = chromPeakTable_ms2)
plot_chrDfList(mchrDfList)
mchrDfList_new <- filterChrDf(chrDfList = mchrDfList, weight_rt = 0.5, weight_shape = 0.5,st = 0.70)
plot_chrDfList(chrDfList = mchrDfList_new)
}
