% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrDf.R
\name{generate_chrDf}
\alias{generate_chrDf}
\title{generate_chrDf}
\usage{
generate_chrDf(ndata, cpid = NA, smooth = TRUE, size = 3)
}
\arguments{
\item{ndata}{nth sample data.}

\item{cpid}{chromatographic peak id.}

\item{smooth}{whether or not smooth.}

\item{size}{smoothMean size.}

\item{noise}{noise value.}
}
\value{
chrDf.
}
\description{
Generate a chrDf.
}
\examples{
load("./test_data/swath_data.RData")
chromPeakTable <- dplyr::as_tibble(cbind(xcms::chromPeaks(swath_data),
                                         xcms::chromPeakData(swath_data)),
                                   rownames = "cpid")
chromPeakTable_ms1 <- chromPeakTable \%>\%
  dplyr::filter(ms_level == 1)
chromPeakTable_ms2 <- chromPeakTable \%>\%
  dplyr::filter(ms_level == 2)
chrDf_i <- generate_chrDf(ndata = swath_data, cpid = "CP34", smooth = TRUE, size = 3)
}
