% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrDf.R
\name{generate_chrDfList}
\alias{generate_chrDfList}
\title{generate_chrDfList}
\usage{
generate_chrDfList(
  ndata,
  cpid,
  chromPeakTable_ms1,
  chromPeakTable_ms2,
  noise = 10,
  smooth = TRUE,
  size = 3,
  factor = 0.5
)
}
\arguments{
\item{ndata}{nth data.}

\item{cpid}{chromatographic peak id.}

\item{chromPeakTable_ms1}{1 level chromPeakTable which have been filtered.(must be generate by ndata)}

\item{chromPeakTable_ms2}{2 level chromPeakTable which have been filtered.(must be generate by ndata)}

\item{noise}{noise value.}

\item{smooth}{whether or not smooth.}

\item{size}{smoothMean size.}

\item{factor}{factor * peakwidth -> rtRange.}
}
\value{
A chrDf.
}
\description{
Generate a chrDfList, contains ms1 chrDf and associated ms2 chrDf. And chromatographic peak will be
detect from chromPeakTable_ms1 and chromPeakTable_ms2.
}
\examples{
load("./test_data/swath_data.RData")
chromPeakTable <- dplyr::as_tibble(cbind(xcms::chromPeaks(swath_data),
                                        xcms::chromPeakData(swath_data)),
                                  rownames = "cpid")
chromPeakTable_ms1 <- chromPeakTable \%>\%
 dplyr::filter(ms_level == 1)
chromPeakTable_ms2 <- chromPeakTable \%>\%
 dplyr::filter(ms_level == 2)
# filter chromPeakTable
chromPeakTable_ms1 <- chromPeakTable_ms1 \%>\%
  dplyr::filter(maxo >= 1000)
chromPeakTable_ms2 <- chromPeakTable_ms2 \%>\%
  dplyr::filter(maxo > 100)
chrDfList <- generate_chrDfList(ndata = swath_data, cpid = "CP34",
                                chromPeakTable_ms1 = chromPeakTable_ms1, chromPeakTable_ms2 = chromPeakTable_ms2)
plot_chrDfList(chrDfList)
}
