% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_chrDf}
\alias{get_chrDf}
\title{get_chrDf}
\usage{
get_chrDf(ndata, m)
}
\arguments{
\item{ndata}{nth sample data.}

\item{m}{mth feature}
}
\value{
A list contains chrDf which belongs to mth ms1 feature.
}
\description{
Get all related ms2 features chromatogram information of the mth ms1 feature.
}
\examples{
load("./test_data/swath_data.RData")
get_chrDf(ndata = swath_data, m = 34)
}
