% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{get_chrDfList}
\alias{get_chrDfList}
\title{get_chrDfList}
\usage{
get_chrDfList(
  ndata,
  m,
  smooth = TRUE,
  noise = 10,
  factor = 1.5,
  sn = 10,
  maxo = 1000
)
}
\arguments{
\item{ndata}{nth sample data.}

\item{m}{mth feature.}

\item{smooth}{Smoothing or not.}

\item{noise}{noise.}

\item{factor}{factor * peakwidth -> rtRange}

\item{sn}{sn threshold.}
}
\value{
A list contains chrDf which belongs to mth ms1 feature.
}
\description{
Get all related ms2 features chromatogram information of the mth ms1 feature.
}
\examples{
load("./test_data/swath_data.RData")
mchrDfList <- get_chrDfList(ndata = swath_data, m = 34)
}
