% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{sp2spMat}
\alias{sp2spMat}
\title{sp2spMat}
\usage{
sp2spMat(sp)
}
\arguments{
\item{sp}{A Spectra object}
}
\value{
A spMat
}
\description{
Transformer a Spectra object to spMat.
}
\examples{
sp2spMat(sp = sp_ms2)
}
