% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_denoising.R
\name{electronic_denoising}
\alias{electronic_denoising}
\title{Spectral Denoising}
\usage{
electronic_denoising(msms)
}
\arguments{
\item{msms}{\code{matrix()} with mz and intensity}
}
\value{
\code{matrix()} with mz and intensity
}
\description{
Noise reduction for MS/MS, including electronic noise and chemical noise.
Electronic denoising method refers from reference (Denoising Search doubles the number of metabolite and exposome annotations in human plasma using an Orbitrap Astral mass spectrometer)
Chemical denoising method is similar with MS2Purifier
}
\examples{
set.seed(110)
msms_data <- .simulate_msms()
result_matrix <- electronic_denoising(msms_data)

plot(msms_data[,1], msms_data[,2], type = "h",
     main = "原始质谱数据 (含电子噪音)",
     xlab = "m/z", ylab = "Intensity")
plot(result_matrix[, "mz"], result_matrix[, "intensity"], type = "h",
     main = "去噪后的质谱数据 (矩阵格式)",
     xlab = "m/z", ylab = "Intensity")
}
