% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{spMat2sp}
\alias{spMat2sp}
\title{spMat2sp}
\usage{
spMat2sp(spMat, msLevel = 1L, rtime = 0)
}
\arguments{
\item{spMat}{A spMat.}
}
\value{
A Spectra object.
}
\description{
Transfor spMat 2 sp.
}
\examples{
spMat2sp(sp_ms2_spMat, msLevel = 2L, rtime = 20)
}
