% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{Deconv4ndata}
\alias{Deconv4ndata}
\title{Deconv4ndata}
\usage{
Deconv4ndata(
  ndata,
  chromPeakTable_ms1,
  chromPeakTable_ms2,
  noise = 10,
  smooth = TRUE,
  size = 3,
  factor = 0.5,
  weight_rt = 0.5,
  weight_shape = 0.5,
  st = 0.8,
  thread = 1
)
}
\arguments{
\item{ndata}{nth sample data.}

\item{chromPeakTable_ms1}{1 level chromPeakTable which have been filtered.(must be generate by ndata)}

\item{chromPeakTable_ms2}{2 level chromPeakTable which have been filtered.(must be generate by ndata)}

\item{noise}{noise value.}

\item{smooth}{whether or not smooth.}

\item{size}{smoothMean size.}

\item{factor}{factor * peakwidth -> rtRange.}

\item{weight_rt}{weight_rt.}

\item{weight_shape}{weight_shape.}

\item{st}{socre threshold.}

\item{thread}{prarllel thread.}
}
\value{
A new featureTable_ms1 with ms2 spectra.
}
\description{
Deconv function for nth sample.
}
\examples{
load("./test_data/swath_data.RData")
chromPeakTable <- dplyr::as_tibble(cbind(xcms::chromPeaks(swath_data),
                                        xcms::chromPeakData(swath_data)),
                                  rownames = "cpid")
chromPeakTable_ms1 <- chromPeakTable \%>\%
 dplyr::filter(ms_level == 1)
chromPeakTable_ms2 <- chromPeakTable \%>\%
 dplyr::filter(ms_level == 2)
# filter chromPeakTable
chromPeakTable_ms1 <- chromPeakTable_ms1 \%>\%
  dplyr::filter(maxo >= 1000)
chromPeakTable_ms2 <- chromPeakTable_ms2 \%>\%
  dplyr::filter(maxo > 100)
chromPeakTable_ms1 <- Deconv4ndata(ndata = swath_data, chromPeakTable_ms1 = chromPeakTable_ms1, chromPeakTable_ms2 = chromPeakTable_ms2)
}
