% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peaks.R
\name{extract_ms1Peaks}
\alias{extract_ms1Peaks}
\alias{extract_ms2Peaks}
\title{Extract ms1 or ms2 peaks}
\usage{
extract_ms1Peaks(
  targetMz,
  dps_ms1,
  scanIndex_ms1,
  rtime_ms1,
  ppm = 5,
  peakwidth = c(5, 20),
  snthresh = 3,
  noise = 100,
  r2thresh = 0.8,
  peakWidthExtend = 1
)

extract_ms2Peaks(
  ms1PeakDT,
  fragmentMz,
  dps_ms2,
  scanIndex_ms2,
  rtime_ms2,
  ppm = 5,
  peakwidth = c(5, 20),
  snthresh = 3,
  noise = 100,
  r2thresh = 0.8
)
}
\arguments{
\item{targetMz}{target m/z}

\item{dps_ms1}{dps_ms1}

\item{scanIndex_ms1}{scanIndex ms1}

\item{rtime_ms1}{rtime ms1}

\item{ppm}{ppm}

\item{peakwidth}{peakwidth}

\item{snthresh}{snthresh}

\item{noise}{noise}

\item{r2thresh}{r2thresh}

\item{peakWidthExtend}{peakWidthExtend}

\item{ms1PeakDT}{ms1PeakDT}

\item{fragmentMz}{fragmentMz}

\item{dps_ms2}{dps_ms2}

\item{scanIndex_ms2}{scanIndex_ms2}

\item{rtime_ms2}{rtime_ms2}
}
\value{
ms1PeaksDT
}
\description{
Extract ms1 peaks from target mz or extract ms2 peaks from ms1PeaksDT
}
