% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{filterChrDf}
\alias{filterChrDf}
\title{filterChrDf}
\usage{
filterChrDf(chrDfList, weight_rt = 0.4, weight_shape = 0.6, st = 0.8)
}
\arguments{
\item{chrDfList}{A chrDfList contains ms1 and ms2 chrDf}
}
\value{
A new chrDfList.
}
\description{
Filter less relevant ms2 features
}
\examples{
mchrDfList <- get_chrDfList(ndata = swath_data, m = 38)
plot_chrDfList(chrDfList = mchrDfList)
mchrDfList_new <- filterChrDf(chrDfList = mchrDfList, weight_rt = 0.4, weight_shape = 0.6,st = 0.80)
plot_chrDfList(chrDfList = mchrDfList_new)
}
