% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_chrDfList}
\alias{get_chrDfList}
\title{get_chrDfList}
\usage{
get_chrDfList(ndata, m, smooth = TRUE)
}
\arguments{
\item{ndata}{nth sample data.}

\item{m}{mth feature.}

\item{smooth}{Smoothing or not.}
}
\value{
A list contains chrDf which belongs to mth ms1 feature.
}
\description{
Get all related ms2 features chromatogram information of the mth ms1 feature.
}
\examples{
load("./test_data/swath_data.RData")
mchrDfList <- get_chrDfList(ndata = swath_data, m = 34)
}
